+++ src/module/users/repository/userRepository.js
@@ -1,5 +1,7 @@
 const { fromDbToEntity } = require('../mapper/userMapper')
 const AbstractUserRepository = require('./abstractUserRepository')
+const UserIdNotDefinedError = require('../controller/error/userIdNotDefinedError')
+const UserNotDefinedError = require('../controller/error/userNotDefinedError')
 
 module.exports = class UserRepository extends AbstractUserRepository {
 
@@ -19,12 +21,17 @@ module.exports = class UserRepository extends AbstractUserRepository {
     }
 
     async getById(id) {
-        console.log(id)
+        if (!id) {
+            throw new UserIdNotDefinedError()
+        }
         const user = await this.userModel.findByPk(id)
         return fromDbToEntity(user)
     }
 
     async remove(id) {
+        if (!id) {
+            throw new UserIdNotDefinedError()
+        }
         const user = await this.userModel.destroy({
             where: {
                 id: id
@@ -34,6 +41,9 @@ module.exports = class UserRepository extends AbstractUserRepository {
     }
 
     async save(user) {
+        if (!user) {
+            throw new UserNotDefinedError()
+        }
         let newUser
         const options = { isNewRecord: !user.id }
         newUser = this.userModel.build(user, options)
